requirements.txt
accelerate==0.27.2
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aioprometheus==23.12.0
aiosignal==1.3.1
anyio==4.5.2
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
beniget==0.4.2.post1
certifi==2023.11.17
charset-normalizer==3.4.1
click==8.1.8
codebleu==0.6.1
contourpy==1.1.1
cycler==0.12.1
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.115.6
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
GPUtil==1.4.0
graphviz==0.20.3
h11==0.14.0
httptools==0.6.4
huggingface-hub==0.17.3
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.5
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
litellm==1.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.1
ninja==1.11.1.3
numpy==1.24.4
openai==0.28.0
orjson==3.10.14
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
propcache==0.2.0
protobuf==5.29.3
psutil==6.1.1
pyarrow==17.0.0
pydantic==1.10.13
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
quantile-python==1.1
ray==2.10.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.20.1
safetensors==0.5.2
scipy==1.10.1
seaborn==0.13.2
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
stdlib-list==0.10.0
sympy==1.13.3
tiktoken==0.7.0
tokenizers==0.14.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
tree-sitter==0.20.4
triton==2.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.21.0
venn==0.1.3
venny4py==1.0.3
vllm==0.2.4
watchfiles==0.24.0
websockets==13.1
xformers==0.0.22.post7
yarl==1.15.2
zipp==3.20.2
